Adicionando service e restcontroller da classe Produto

+++ java/seboDoJoao/src/main/java/br/com/interdisciplinar/sebodojoao/repository/ProdutoRepository.java
@@ -3,6 +3,5 @@ package br.com.interdisciplinar.sebodojoao.repository;
 import br.com.interdisciplinar.sebodojoao.model.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
-
+public interface ProdutoRepository extends JpaRepository<Produto, Long> {
 }
